+++ packages/integrations/markdoc/test/render-html.test.js
@@ -253,74 +253,3 @@ function assertInlineMark(el) {
   expect(el.className).to.equal('mark');
   expect(el.style.color).to.equal('hotpink');
 }
-
-// for reference, here's what the component rendered HTML looks like
-/* 
-<article>
-  <p>This is a <span style="color: hotpink" class="mark">inline mark</span> in regular Markdown markup.</p>
-  <p id="p1">This is a <span style="color: hotpink" class="mark">inline mark</span> under some HTML</p>
-  <div id="div1">
-    <p id="p2">This is a <span style="color: hotpink" class="mark">inline mark</span> under some HTML</p>
-    <p id="p3">This is a <span id="p3-span1"><span style="color: hotpink" class="mark">inline mark</span></span> under
-      some HTML</p>
-  </div>
-  <aside class="content tip astro-DUQFCLOB" aria-label="Aside One">
-    <p class="title astro-DUQFCLOB" aria-hidden="true">
-      <span class="icon astro-DUQFCLOB">
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 18 18" width="16" height="16" class="astro-DUQFCLOB">
-          <path fill-rule="evenodd"
-            d="M14 0a8.8 8.8 0 0 0-6 2.6l-.5.4-.9 1H3.3a1.8 1.8 0 0 0-1.5.8L.1 7.6a.8.8 0 0 0 .4 1.1l3.1 1 .2.1 2.4 2.4.1.2 1 3a.8.8 0 0 0 1 .5l2.9-1.7a1.8 1.8 0 0 0 .8-1.5V9.5l1-1 .4-.4A8.8 8.8 0 0 0 16 2v-.1A1.8 1.8 0 0 0 14.2 0h-.1zm-3.5 10.6-.3.2L8 12.3l.5 1.8 2-1.2a.3.3 0 0 0 .1-.2v-2zM3.7 8.1l1.5-2.3.2-.3h-2a.3.3 0 0 0-.3.1l-1.2 2 1.8.5zm5.2-4.5a7.3 7.3 0 0 1 5.2-2.1h.1a.3.3 0 0 1 .3.3v.1a7.3 7.3 0 0 1-2.1 5.2l-.5.4a15.2 15.2 0 0 1-2.5 2L7.1 11 5 9l1.5-2.3a15.3 15.3 0 0 1 2-2.5l.4-.5zM12 5a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-8.4 9.6a1.5 1.5 0 1 0-2.2-2.2 7 7 0 0 0-1.1 3 .2.2 0 0 0 .3.3c.6 0 2.2-.4 3-1.1z"
-            class="astro-DUQFCLOB"></path>
-        </svg>
-      </span>
-      Aside One
-    </p>
-    <section class="astro-DUQFCLOB">
-      <p>I&#39;m a Markdown paragraph inside an top-level aside tag</p>
-      <h2 id="im-an-h2-via-markdown-markup">I&#39;m an H2 via Markdown markup</h2>
-      <h2 id="h-two">I&#39;m an H2 via HTML markup</h2>
-      <p><strong>Markdown bold</strong> vs <strong>HTML bold</strong></p>
-      <p></p>
-    </section>
-  </aside>
-  <section id="section1">
-    <div id="div1">
-      <aside class="content tip astro-DUQFCLOB" aria-label="Nested un-indented Aside">
-        <p class="title astro-DUQFCLOB" aria-hidden="true">
-          <span class="icon astro-DUQFCLOB">
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 18 18" width="16" height="16" class="astro-DUQFCLOB">
-              <path fill-rule="evenodd"
-                d="M14 0a8.8 8.8 0 0 0-6 2.6l-.5.4-.9 1H3.3a1.8 1.8 0 0 0-1.5.8L.1 7.6a.8.8 0 0 0 .4 1.1l3.1 1 .2.1 2.4 2.4.1.2 1 3a.8.8 0 0 0 1 .5l2.9-1.7a1.8 1.8 0 0 0 .8-1.5V9.5l1-1 .4-.4A8.8 8.8 0 0 0 16 2v-.1A1.8 1.8 0 0 0 14.2 0h-.1zm-3.5 10.6-.3.2L8 12.3l.5 1.8 2-1.2a.3.3 0 0 0 .1-.2v-2zM3.7 8.1l1.5-2.3.2-.3h-2a.3.3 0 0 0-.3.1l-1.2 2 1.8.5zm5.2-4.5a7.3 7.3 0 0 1 5.2-2.1h.1a.3.3 0 0 1 .3.3v.1a7.3 7.3 0 0 1-2.1 5.2l-.5.4a15.2 15.2 0 0 1-2.5 2L7.1 11 5 9l1.5-2.3a15.3 15.3 0 0 1 2-2.5l.4-.5zM12 5a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-8.4 9.6a1.5 1.5 0 1 0-2.2-2.2 7 7 0 0 0-1.1 3 .2.2 0 0 0 .3.3c.6 0 2.2-.4 3-1.1z"
-                class="astro-DUQFCLOB"></path>
-            </svg>
-          </span>
-          Nested un-indented Aside
-        </p>
-        <section class="astro-DUQFCLOB">
-          <p>regular Markdown markup</p>
-          <p id="p4">nested <span style="color: hotpink" class="mark">inline mark</span> content</p>
-        </section>
-      </aside>
-    </div>
-    <div id="div2">
-      <aside class="content tip astro-DUQFCLOB" aria-label="Nested indented Aside">
-        <p class="title astro-DUQFCLOB" aria-hidden="true">
-          <span class="icon astro-DUQFCLOB">
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 18 18" width="16" height="16" class="astro-DUQFCLOB">
-              <path fill-rule="evenodd"
-                d="M14 0a8.8 8.8 0 0 0-6 2.6l-.5.4-.9 1H3.3a1.8 1.8 0 0 0-1.5.8L.1 7.6a.8.8 0 0 0 .4 1.1l3.1 1 .2.1 2.4 2.4.1.2 1 3a.8.8 0 0 0 1 .5l2.9-1.7a1.8 1.8 0 0 0 .8-1.5V9.5l1-1 .4-.4A8.8 8.8 0 0 0 16 2v-.1A1.8 1.8 0 0 0 14.2 0h-.1zm-3.5 10.6-.3.2L8 12.3l.5 1.8 2-1.2a.3.3 0 0 0 .1-.2v-2zM3.7 8.1l1.5-2.3.2-.3h-2a.3.3 0 0 0-.3.1l-1.2 2 1.8.5zm5.2-4.5a7.3 7.3 0 0 1 5.2-2.1h.1a.3.3 0 0 1 .3.3v.1a7.3 7.3 0 0 1-2.1 5.2l-.5.4a15.2 15.2 0 0 1-2.5 2L7.1 11 5 9l1.5-2.3a15.3 15.3 0 0 1 2-2.5l.4-.5zM12 5a1 1 0 1 1-2 0 1 1 0 0 1 2 0zm-8.4 9.6a1.5 1.5 0 1 0-2.2-2.2 7 7 0 0 0-1.1 3 .2.2 0 0 0 .3.3c.6 0 2.2-.4 3-1.1z"
-                class="astro-DUQFCLOB"></path>
-            </svg>
-          </span>
-          Nested indented Aside
-        </p>
-        <section class="astro-DUQFCLOB">
-          <p>regular Markdown markup</p>
-          <p id="p5">nested <span style="color: hotpink" class="mark">inline<span
-                id="inception-span">mark</span></span> content</p>
-        </section>
-      </aside>
-    </div>
-  </section>
-</article>
-*/
